[infra] Restructure & initial work on requests
