+++ pkg/infra/game/agent/agent.go
@@ -1,7 +1,6 @@
 package agent
 
 import (
-	"fmt"
 	"infra/game/commons"
 	"infra/game/decision"
 	"infra/game/message"
@@ -13,8 +12,9 @@ import (
 )
 
 type Strategy interface {
-	HandleFightMessage(m message.TaggedMessage, view *state.View, agent BaseAgent, log *immutable.Map[commons.ID, decision.FightAction]) decision.FightAction
-	Default() decision.FightAction
+	HandleFightInformation(m message.TaggedMessage, view *state.View, agent BaseAgent, log *immutable.Map[commons.ID, decision.FightAction])
+	HandleFightRequest(m message.TaggedMessage, view *state.View, log *immutable.Map[commons.ID, decision.FightAction]) message.Payload
+	CurrentAction() decision.FightAction
 }
 
 type Agent struct {
@@ -24,7 +24,8 @@ type Agent struct {
 
 func (a *Agent) HandleFight(view state.View, log immutable.Map[commons.ID, decision.FightAction], decisionChan chan message.ActionMessage, wg *sync.WaitGroup) {
 	for m := range a.BaseAgent.communication.receipt {
-		action := a.handleMessage(&view, &log, m)
+		a.handleMessage(&view, &log, m)
+		action := a.Strategy.CurrentAction()
 		if action != decision.Undecided {
 			go func() {
 				<-a.BaseAgent.communication.receipt
@@ -34,76 +35,20 @@ func (a *Agent) HandleFight(view state.View, log immutable.Map[commons.ID, decis
 			return
 		}
 	}
-	decisionChan <- message.ActionMessage{Action: a.Strategy.Default(), Sender: a.BaseAgent.Id}
+	decisionChan <- message.ActionMessage{Action: a.Strategy.CurrentAction(), Sender: a.BaseAgent.Id}
 }
 
 func (a *Agent) handleMessage(view *state.View, log *immutable.Map[commons.ID, decision.FightAction], m message.TaggedMessage) decision.FightAction {
 	switch m.Message.MType() {
 	case message.Close:
+	case message.Request:
+		payload := a.Strategy.HandleFightRequest(m, view, log)
+		err := a.BaseAgent.SendBlockingMessage(m.Sender, *message.NewMessage(message.Inform, payload))
+		logging.Log(logging.Error, nil, err.Error())
+	case message.Inform:
+		a.Strategy.HandleFightInformation(m, view, a.BaseAgent, log)
 	default:
-		fightMessage := a.Strategy.HandleFightMessage(m, view, a.BaseAgent, log)
-		return fightMessage
+		a.Strategy.HandleFightInformation(m, view, a.BaseAgent, log)
 	}
-	return decision.Undecided
-}
-
-func (ba *BaseAgent) log(lvl logging.Level, fields logging.LogField, msg string) {
-	agentFields := logging.LogField{
-		"agentName": ba.AgentName,
-		"agentID":   ba.Id,
-	}
-
-	logging.Log(lvl, logging.CombineFields(agentFields, fields), msg)
-}
-
-type BaseAgent struct {
-	communication Communication
-	Id            commons.ID
-	AgentName     string
-}
-
-func NewBaseAgent(communication Communication, id commons.ID, agentName string) BaseAgent {
-	return BaseAgent{communication: communication, Id: id, AgentName: agentName}
-}
-
-type Communication struct {
-	receipt <-chan message.TaggedMessage
-	peer    immutable.Map[commons.ID, chan<- message.TaggedMessage]
-}
-
-func NewCommunication(receipt <-chan message.TaggedMessage, peer immutable.Map[commons.ID, chan<- message.TaggedMessage]) Communication {
-	return Communication{receipt: receipt, peer: peer}
-}
-
-func (b BaseAgent) broadcastBlockingMessage(m message.Message) {
-	iterator := b.communication.peer.Iterator()
-	tm := message.TaggedMessage{
-		Sender:  b.Id,
-		Message: m,
-	}
-	for !iterator.Done() {
-		_, c, ok := iterator.Next()
-		if ok {
-			c <- tm
-		}
-	}
-}
-
-func (b BaseAgent) sendBlockingMessage(id commons.ID, m message.Message) (e error) {
-	defer func() {
-		if r := recover(); r != nil {
-			e = fmt.Errorf("agent %s not available for messaging, submitted", id)
-		}
-	}()
-
-	value, ok := b.communication.peer.Get(id)
-	if ok {
-		value <- message.TaggedMessage{
-			Sender:  b.Id,
-			Message: m,
-		}
-	} else {
-		e = fmt.Errorf("agent %s not available for messaging, dead", id)
-	}
-	return
+	return a.Strategy.CurrentAction()
 }
