+++ pkg/infra/game/message/message.go
@@ -1,6 +1,7 @@
 package message
 
 import (
+	"github.com/google/uuid"
 	"infra/game/commons"
 	"infra/game/decision"
 )
@@ -13,8 +14,8 @@ type Type int64
 
 const (
 	Close Type = iota
-	Something
-	SomethingElse
+	Request
+	Inform
 )
 
 type Message struct {
@@ -26,17 +27,18 @@ func NewMessage(mType Type, payload Payload) *Message {
 	return &Message{mType: mType, payload: payload}
 }
 
-func (m Message) MType() Type {
+func (m *Message) MType() Type {
 	return m.mType
 }
 
-func (m Message) Payload() Payload {
+func (m *Message) Payload() Payload {
 	return m.payload
 }
 
 type TaggedMessage struct {
 	Sender  commons.ID
 	Message Message
+	mId     uuid.UUID
 }
 
 type ActionMessage struct {
